+++ noodles-csi/src/io.rs
@@ -1,0 +1,5 @@
+//! CSI I/O.
+
+mod query;
+
+pub use self::query::Query;